add : component
studentPage : add GuideButton, LinkBox, NavBar component
teacherPage : add Korea, Incheon, NavBar, Notice, StudentList component
